deleted dbConnection.js and added connection code into index.js
