+++ index.js
@@ -1,8 +1,28 @@
 const inquirer = require("inquirer");
 const cTable = require("console.table");
-const dbConnection = require("./sql/dbConnection.js");
 const fs = require("fs");
 const path = require("path");
+const mysql = require("mysql");
+// establish server connection configuration
+var connection = mysql.createConnection({
+	host: "localhost",
+
+	// my local MySQL default port
+	port: 3306,
+
+	// username
+	user: "root",
+
+	// password
+	password: "server_password",
+	database: "employee_db",
+});
+// creating the connection to the employee_db database
+connection.connect(function (err) {
+	if (err) throw err;
+	console.log("connected as id " + connection.threadId);
+	connection.end();
+});
 
 // Function to validaate that each questions is entered.
 function validation() {
@@ -114,10 +134,7 @@ const updateEmployeeRoleQuestion = [
 
 // Function to view all employees
 function viewAllEmployees() {
-	dbConnection.connection.query("CALL view_all_employees()", function (
-		err,
-		res
-	) {
+	connection.query("CALL view_all_employees()", function (err, res) {
 		if (err) throw err;
 		console.table(res);
 		connection.end();
